move component into components
